feat: add uuidv4, swagger in archivos

+++ src/config/database/migrations/20240924000003-insert-roles.ts
@@ -1,15 +1,27 @@
 'use strict';
 
-import { QueryInterface, Sequelize } from 'sequelize';
+import { QueryInterface } from 'sequelize';
+import { v4 as uuidv4 } from 'uuid';
 
 module.exports = {
-  up: async (queryInterface: QueryInterface, Sequelize: Sequelize) => {
+  up: async (queryInterface: QueryInterface) => {
     await queryInterface.bulkInsert('Rol', [
-      { descripcion: 'admin', createdAt: new Date(), updatedAt: new Date() },
-      { descripcion: 'productor', createdAt: new Date(), updatedAt: new Date() },
+      {
+        id_rol: uuidv4(),
+        descripcion: 'admin',
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      },
+      {
+        id_rol: uuidv4(),
+        descripcion: 'productor',
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      },
     ]);
   },
-  down: async (queryInterface: QueryInterface, Sequelize: Sequelize) => {
-    await queryInterface.bulkDelete('Rol', {}, {});
+
+  down: async (queryInterface: QueryInterface) => {
+    await queryInterface.bulkDelete('Rol', { descripcion: ['admin', 'productor'] });
   },
 };
